requirements.txt
# Core dependencies
numpy<2.0.0  # Required for compatibility with ChromaDB
python-dotenv>=1.0.0
flask>=2.0.0
requests>=2.28.0

# AI/ML dependencies
google-generativeai>=0.8.0
sentence-transformers==2.2.2  # Specific version to ensure compatibility with ChromaDB
chromadb>=0.4.18

# AutoGen dependencies
pyautogen>=0.2.0
openai>=0.28.0  # For OpenAI fallback

# Web dependencies
fastapi>=0.95.2  # Used in ChromaDB
uvicorn>=0.18.3  # Used in ChromaDB
websockets>=10.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.65.0
pytest>=7.3.1

# Web Scraping
beautifulsoup4==4.12.2

# Async processing
asyncio>=3.4.3

# Backend
websockets==11.0.3

